requirements.txt
requests
pyyaml
dotenv
